+++ src/middleware/client.js
@@ -1,5 +1,6 @@
 import _ from "lodash";
 import jwt from "jwt-simple";
+
 import { handleExtract, requestExtract } from "../helpers";
 
 function parseQueryString(query) {
@@ -28,7 +29,11 @@ function parseToken(token, secret) {
 
 
 module.exports = function hullClientMiddlewareFactory(Client, { hostSecret, clientConfig = {} }) {
-  function getCurrentShip(id, client, cache, bust) {
+  function getCurrentShip(id, client, cache, bust, notification) {
+    if (notification && notification.connector) {
+      return Promise.resolve(notification.connector);
+    }
+
     return (() => {
       if (cache && bust) {
         return cache.del(id);
@@ -61,7 +66,7 @@ module.exports = function hullClientMiddlewareFactory(Client, { hostSecret, clie
         parseToken(req.hull.token, hostSecret) ||
         parseQueryString(req.query) ||
         {};
-      const { message, config } = req.hull;
+      const { message, notification, config } = req.hull;
       const { organization, ship: id, secret } = config;
       if (organization && id && secret) {
         req.hull.client = new Client(_.merge({ id, secret, organization }, clientConfig));
@@ -78,8 +83,9 @@ module.exports = function hullClientMiddlewareFactory(Client, { hostSecret, clie
         req.hull.token = jwt.encode(config, hostSecret);
 
         const bust = (message && message.Subject === "ship:update");
+
         // Promise<ship>
-        return getCurrentShip(id, req.hull.client, req.hull.cache, bust).then((ship = {}) => {
+        return getCurrentShip(id, req.hull.client, req.hull.cache, bust, notification).then((ship = {}) => {
           req.hull.ship = ship;
           req.hull.hostname = req.hostname;
           req.hull.options = _.merge(req.query, req.body);