+++ src/utils/index.js
@@ -1,5 +1,6 @@
 export exitHandler from "./exit-handler";
 export notifHandler from "./notif-handler";
+export smartNotifierHandler from "./smart-notifier-handler";
 export oAuthHandler from "./oauth-handler";
 export actionHandler from "./action-handler";
 export batcherHandler from "./batcher-handler";
@@ -10,5 +11,10 @@ export tokenMiddleware from "./token-middleware";
 export requireHullMiddleware from "./require-hull-middleware";
 export responseMiddleware from "./response-middleware";
 export notifMiddleware from "./notif-middleware";
+export smartNotifierMiddleware from "./smart-notifier-middleware";
+export smartNotifierErrorMiddleware from "./smart-notifier-error-middleware";
 export segmentsMiddleware from "./segments-middleware";
 export helpersMiddleware from "./helpers-middleware";
+export SmartNotifierResponse from "./smart-notifier-response";
+export PromiseReuser from "./promise-reuser";
+