+++ test/unit/utils/smart-notifier-middleware-test.js
@@ -1,0 +1,13 @@
+/* global describe, it */
+import { expect, should } from "chai";
+import sinon from "sinon";
+
+import smartNotifierMiddleware from "../../../src/utils/smart-notifier-middleware";
+
+
+describe("SmartNotifierMiddleware", () => {
+  it("should return a middleware function", () => {
+    const testInstance = new smartNotifierMiddleware({});
+    expect(typeof testInstance).to.equal("function");
+  });
+});