+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# 0.13.1
+* adds error handler and segment filtering to `smartNotifierHandler`
+* properly map account object from batch extract to notification format
+* fetch 200 segments for segments middleware
+
+# 0.13.0
+* introduces `smartNotifierMiddleware` and `smartNotifierHandler` utils
+* introduces `ctx.smartNotifierResponse` context object to build flow response
+* added `skipSignatureValidation` `Hull.Connector` options to skip smartNotifier signature validation
+* **breaking**: removed `serviceMiddleware` and `service` param on `Hull.Connector`, new way of applying service related objects to the Context Object is the `use` method on the Connector instance and initialize the service objects directly on the `req.hull.service` namespace
+* restructurized tests folders and files
+
+# 0.12.7
+* makes `PromiseReuser` reuse the whole `cache.wrap`, not `client.get` only
+
+# 0.12.6
+* adds `PromiseReuser` class which allows us to handle in memory longer running promises and reuse them in case the same function is called multiple times with the same input
+
+# 0.12.5
+* upgrades hull-client to 1.1.3
+* update documentation about initialization of `QueueAdapter`
+
+# 0.12.4
+* handle properly empty jobs in background worker
+
+# 0.12.3
+* upgrades underlying hull-client which fixes the way timeouted or errored rest API events are rejected
+
 # 0.12.2
 * adds option to get cache data via `req.hull.cache.get`